Remove Container for KeyMessages and use component
- connect() was causing the component to rerender everytime the state changes from the blockchain listener. This reset the counter.
