+++ src/app/tabs/registrar/helpers.js
@@ -72,3 +72,23 @@ export const getAddrRegisterData = (name, owner, secret, duration, addr) => {
 
   return `${dataSignature}${dataOwner}${dataSecret}${dataDuration}${dataAddr}${dataName}`;
 };
+
+/**
+ * Shuffles an array
+ * @param {array} array to be shuffled
+ */
+export const shuffle = (array) => {
+  let counter = array.length;
+  const returnAray = [];
+
+  while (counter > 0) {
+    const index = Math.floor(Math.random() * counter);
+    counter -= 1;
+
+    const temp = array[counter];
+    returnAray[counter] = array[index];
+    returnAray[index] = temp;
+  }
+
+  return returnAray;
+};
